Cleanup Unneeded WidgetUpdater.kt
